Fixed a couple of errors, improved the BG system, added AlteracValley, added a static path system, centraliced all movements to MovementEngine, improved the targeting system, tested with Druid Resto

+++ AmeisenBotX.Core/Engines/Movement/Settings/MovementSettings.cs
@@ -12,8 +12,8 @@
 
         public float SeperationDistance { get; set; } = 2.0f;
 
-        public double WaypointCheckThreshold { get; set; } = 1.7;
+        public float WaypointCheckThreshold { get; set; } = 1.7f;
 
-        public double WaypointCheckThresholdMounted { get; set; } = 3.5;
+        public float WaypointCheckThresholdMounted { get; set; } = 3.5f;
     }
 }
